+++ persistence/src/androidMain/kotlin/com/wire/kalium/persistence/util/JsonSerializer.kt
@@ -1,0 +1,9 @@
+package com.wire.kalium.persistence.util
+
+import kotlinx.serialization.json.Json
+
+object JsonSerializer  {
+    operator fun invoke() = Json {
+        encodeDefaults = true
+    }
+}
