+++ persistence/src/commonMain/kotlin/com/wire/kalium/persistence/model/Session.kt
@@ -1,0 +1,10 @@
+package com.wire.kalium.persistence.model
+
+data class Session(
+    val userId: String,
+    val clientId: String,
+    val domain: String,
+    val tokenType: String,
+    val accessToken: String,
+    val refreshToken: String
+)
