--- persistent/src/androidMain/AndroidManifest.xml
@@ -1,3 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.wire.kalium.persistent" />
