--- persistent/src/androidMain/kotlin/com/wire/kalium/persistent/util/JsonSerializer.kt
@@ -1,9 +1,0 @@
-package com.wire.kalium.persistent.util
-
-import kotlinx.serialization.json.Json
-
-object JsonSerializer  {
-    operator fun invoke() = Json {
-        encodeDefaults = true
-    }
-}
