--- persistent/src/androidMain/kotlin/com/wire/kalium/persistent/util/SecurityUtil.kt
@@ -1,60 +1,0 @@
-package com.wire.kalium.persistent.util
-
-import android.security.keystore.KeyGenParameterSpec
-import android.security.keystore.KeyProperties.BLOCK_MODE_GCM
-import android.security.keystore.KeyProperties.ENCRYPTION_PADDING_NONE
-import android.security.keystore.KeyProperties.KEY_ALGORITHM_AES
-import android.security.keystore.KeyProperties.PURPOSE_DECRYPT
-import android.security.keystore.KeyProperties.PURPOSE_ENCRYPT
-import java.security.KeyStore
-import javax.crypto.Cipher
-import javax.crypto.KeyGenerator
-import javax.crypto.SecretKey
-import javax.crypto.spec.GCMParameterSpec
-
-class SecurityUtil {
-    private val cipher by lazy {
-        Cipher.getInstance("AES/GCM/NoPadding")
-    }
-    private val charset by lazy {
-        charset("UTF-8")
-    }
-    private val keyStore by lazy {
-        KeyStore.getInstance(provider).apply {
-            load(null)
-        }
-    }
-    private val keyGenerator by lazy {
-        KeyGenerator.getInstance(KEY_ALGORITHM_AES, provider)
-    }
-
-    fun encryptData(keyAlias: String, text: String): ByteArray {
-        cipher.init(Cipher.ENCRYPT_MODE, generateSecretKey(keyAlias))
-        return cipher.doFinal(text.toByteArray(charset))
-    }
-
-    fun decryptData(keyAlias: String, encryptedData: ByteArray): String {
-        cipher.init(Cipher.DECRYPT_MODE, getSecretKey(keyAlias), GCMParameterSpec(128, cipher.iv))
-        return cipher.doFinal(encryptedData).toString(charset)
-    }
-
-    private fun generateSecretKey(keyAlias: String): SecretKey {
-        return keyGenerator.apply {
-            init(
-                KeyGenParameterSpec
-                    .Builder(keyAlias, PURPOSE_ENCRYPT or PURPOSE_DECRYPT)
-                    .setBlockModes(BLOCK_MODE_GCM)
-                    .setEncryptionPaddings(ENCRYPTION_PADDING_NONE)
-                    .build()
-            )
-        }.generateKey()
-    }
-
-    private fun getSecretKey(keyAlias: String): SecretKey? {
-        return (keyStore.getEntry(keyAlias, null) as KeyStore.SecretKeyEntry).secretKey
-    }
-
-    private companion object {
-        private const val provider = "AndroidKeyStore"
-    }
-}
