--- persistent/src/commonMain/kotlin/com/wire/kalium/persistent/model/Session.kt
@@ -1,10 +1,0 @@
-package com.wire.kalium.persistent.model
-
-data class Session(
-    val userId: String,
-    val clientId: String,
-    val domain: String,
-    val tokenType: String,
-    val accessToken: String,
-    val refreshToken: String
-)
